[Edit] MainView 폰트 및 스페이싱 수정
