+++ FinalHrmiProjects/Resource/Font +.swift
@@ -10,6 +10,7 @@ import SwiftUI
 // MARK: - Font 커스텀 [ pretendard ]
 extension Font {
     // Bold
+    static let bold22: Font = .custom(FontType.Bold.name, size: 22)
     static let bold20: Font = .custom(FontType.Bold.name, size: 20)
     static let bold18: Font = .custom(FontType.Bold.name, size: 18)
     static let bold16: Font = .custom(FontType.Bold.name, size: 16)
